tests/submission/PhilipTaylor/canvas/2d.path.arcTo.shape.curve2.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.path.arcTo.shape.curve2</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.path.html">path</a>.<a href="index.2d.path.arcTo.html">arcTo</a>.<a href="index.2d.path.arcTo.shape.html">shape</a>.curve2</h1>
<p class="desc">arcTo() curves in the right kind of shape</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.path.arcTo.shape">2d.path.arcTo.shape</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="green-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("arcTo() curves in the right kind of shape");
_addTest(function(canvas, ctx) {

var tol = 1.5; // tolerance to avoid antialiasing artifacts

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);

ctx.fillStyle = '#f00';
ctx.beginPath();
ctx.rect(10, 20, 45, 10);
ctx.moveTo(80, 45);
ctx.arc(55, 45, 25-tol, 0, -Math.PI/2, true);
ctx.arc(55, 45, 15+tol, -Math.PI/2, 0, false);
ctx.fill();

ctx.strokeStyle = '#0f0';
ctx.lineWidth = 10;
ctx.beginPath();
ctx.moveTo(10, 25);
ctx.arcTo(75, 25, 75, 60, 20);
ctx.stroke();

_assertPixel(canvas, 50,25, 0,255,0,255, "50,25", "0,255,0,255");
_assertPixel(canvas, 55,19, 0,255,0,255, "55,19", "0,255,0,255");
_assertPixel(canvas, 55,20, 0,255,0,255, "55,20", "0,255,0,255");
_assertPixel(canvas, 55,21, 0,255,0,255, "55,21", "0,255,0,255");
_assertPixel(canvas, 64,22, 0,255,0,255, "64,22", "0,255,0,255");
_assertPixel(canvas, 65,21, 0,255,0,255, "65,21", "0,255,0,255");
_assertPixel(canvas, 72,28, 0,255,0,255, "72,28", "0,255,0,255");
_assertPixel(canvas, 73,27, 0,255,0,255, "73,27", "0,255,0,255");
_assertPixel(canvas, 78,36, 0,255,0,255, "78,36", "0,255,0,255");
_assertPixel(canvas, 79,35, 0,255,0,255, "79,35", "0,255,0,255");
_assertPixel(canvas, 80,44, 0,255,0,255, "80,44", "0,255,0,255");
_assertPixel(canvas, 80,45, 0,255,0,255, "80,45", "0,255,0,255");
_assertPixel(canvas, 80,46, 0,255,0,255, "80,46", "0,255,0,255");


});
</script>

<div id="log"></div>
</body></html>
